publish-techdocs: create reusable workflow for gen+pub techdocs
This is a reusable workflow that installs dependencies, authenticates
against GCP with secret `TECHDOCS_CONTRIB_KEY_JSON` (a static json key
file), installs dependencies, generates the docs, and finally pushes
the generated, static content to the secret `TECHDOCS_BUCKET_NAME`.

It takes input for the namespace, kind and name for the content (used
in storing the content) and also takes in a `default-working-directory`.
This input is used when running `techdocs-cli` commands and is needed
when the `mkdocs.yml` file is not at the root of the project.