Add support for triggering Argo workflows
Here we add a small tool wrapping the `argo` CLI, a Docker image to
build and push it, and an action which uses it.

These let us trigger Argo workflows on our dev and prod instances.
Credentials to trigger are scoped to only be requestable by the listed
repository, and they can only trigger workflows in their own namespace
too. Secrets are not available to forks.