get-vault-secrets: Work with dev or ops vault (#14)
I'm going to make an action to trigger Argo Workflows in our dev
instance. For this, we'll need to be able to fetch secrets from dev
Vault. Update the action to add a new `vault_instance` input which
switches between the two.

Also add a testsuite for this and the existing `bats` tests which we have.