lint-shared-workflows: Also lint action definitions (#110)
These have a schema, so we can check if they're correct beyond just
being valid YAML. Use the JSON schema from schemastore.org to do this -
so we have to convert the YAML files to JSON first. For this, we run in
the `yq` container and use `yq`'s conversion feature.

We cache the schema and run with the cached copy from the FS, to reduce
the chance of interweb flakiness breaking the test.