build-push-to-dockerhub: Generate SBOM & provenance attestations
This generates an SBOM and provenance attestations. These are also
stored in one of GitHub's SigStore instances, so the provenance of
images built using this action can be verified.

We delete the `syft-sbom-report` action. This wasn't a usable action (no
`action.yaml`) anyway.

What we do here is based on [a comment][comment] from one of the GitHub
attestation folks. We only generate the provenance attestation with
GitHub directly, and use Docker to generate the SBOM and another
provenance attestation about the Dockerish bits of the build. The GitHub
attestation acts as a signature over all of that, and should handle
multi-arch manifests properly, which isn't possible with the documented
GitHub way that exists currently.

[comment]: https://github.com/actions/attest-sbom/issues/60#issuecomment-2108917313