lint-shared-workflows: Run for PRs
This is a required check - we need to make sure it runs for pull
requests.

Checking out `${{ github.head_ref }}` from a fork doesn't work either,
so drop that from the `lint` action and use the default. The problem is
that [`pull_request` workflows for forks run in the base
repository][pr-base]. In particular, `${{ github.repository }}` is the
base. So for this commit we were trying to check out `iainlane/pr` from
`grafana/shared-workflows`. But that branch is not on that repository
because it's from a fork: it's on `iainlane/shared-workflows`.

There is a way to check out PRs from only the base repository though, by
using `${{ github.ref }}`. For PRs that is something like
`pull/140/merge`. In fact, using that is the default for
`actions/checkout`, so we'll drop the setting of the `ref` input
altogether.

[pr-base]: https://docs.github.com/en/actions/using-workflows/events-that-trigger-workflows#pull-request-events-for-forked-repositories