CI: Enable CodeQL
I've gone for doing this manually rather than through the GitHub UI.
This is because when I enabled in the UI just now, there were some
results in the generated code `actions/lint-pr-title/dist/index.js`, and
these are better identified by security alerts on the dependencies, if
there is an actual issue. We want to use CodeQL to find our own issues.