ci(test-find-pr-for-commit): Don't run for merge group events (#195)
This action-testing workflow doesn't make sense when run in the merge
group, since the commits there aren't associated with PRs any more. It's
failing.

We could in the future consider doing work to get the association back,
but I don't think we have a need for it right now, so we can stop
running the check there instead.