feat(run-go-program): add action to build/cache/run a go binary
It's common to want to essentially call `go run` on a Go script in CI.
But doing this directly can be slow. While the `setup-go` action does
cache, retrieivng all the dependencies from the cache and then building
the binary each time is still not quick.

Here we're adding a simple action to cache the resulting binary. The
scope is intentionally for simple programs where you only need the
binary you just built to run it, and you can build with `go build`.

If the binary is present in the cache, it is restored. Otherwise, it is
built and cahced for next time.