fix(renovate): fix action versions, add more triggers (#249)
When referencing actions, we'll do:

```yaml
uses: foo/bar-action@123456789 # v1.2.3
```

to reference by SHA but still say what the version was. Dependabot
supports updating these comments when creating bumps, but Renovate
doesn't do this when the comment points to the wrong version.

We had two cases of that in `codeql-action` and `vault-action`. Here we
fix both.

Also, it's easier to test if we can kick off runs manually and if
Renovate triggers when its config or workflow are changed, so add
triggers for that.

[preset]: https://docs.renovatebot.com/presets-helpers/#helperspingithubactiondigeststosemver