feat(trigger-argo-workflow): support {dev,ops}-aws instances and get repo-specifc paths
We're migrating our Argo Workflows environment internally, and for that
we currently have parallel deployments. We've namespaced the secrets
needed to trigger in each cluster. This PR modifies to allow the new
`dev-aws` instance and fetch the cluster-scoped secrets.

To save time for the production migration later, we'll also teach
everything about `ops-aws` now, even though it doesn't exist yet.