fix(ci): run required checks when undrafting PRs (#428)
We just merged `release-please` support in
023c802ae4fc29d9bc2b3428333a63f47a5dc255. The user this runs as is the
default Actions bot user. One thing about this user in particular is
that it doesn't trigger workflows to run.

This means that its release PRs aren't actually mergeable, since we have
a couple of required checks which come from Actions runs.

Fortunately we don't have to switch to an app. We can use the fact that
the PRs are created in draft mode. To make a release, the process will
be

1. Make PR ready for review.
2. Add PR to merge queue.

Workflows can be triggered by the `ready_for_review` activity type on
the `pull_request` event. So they will be kicked off when a human does
step 1.