ci(renovate): fix path to config, update version of renovate itself
One of the parameters to the `renovatebot/github-action` action is the
version of Renovate to use, in the form of a Docker image reference
(version/sha part). This can't be updated by any built in updater, since
it's an arbitrary parameter to an action, but we can use the 'custom
matchers' feature of Renovate to do this. Using that, we can add a
comment above the field which needs to be edited, saying what image the
string refers to, and then Renovate can update it like normal.

While we're here, the paths in this file to the Renovate config were
wrong - fix them.