feat(check-drone-sig): init reusable workflow
This reusable workflow is intended to be used by public, trusted
drone-enabled repos to ensure changes to drone config are properly
signed when the config it updated. Without this check, users can commit
changes to the config without resigning the config file and cause builds
to require manual approval to run.