feat(check-drone-sig): init reusable workflow (#488)
* feat(check-drone-sig): init reusable workflow

This reusable workflow is intended to be used by public, trusted
drone-enabled repos to ensure changes to drone config are properly
signed when the config it updated. Without this check, users can commit
changes to the config without resigning the config file and cause builds
to require manual approval to run.

* Run reformatter

* Remove `required` from inputs

* Use `--quiet` for `git diff`

* Update failure message

* Don't print diff.

* Pin drone cli version