chore(renovate): use json5 config (#511)
Renovate supports JSON or [JSON5]. JSON5 supports comments. For an
upcoming change (adding a custom schedule for updates to Renovate itself
so they happen less often) I want to add a comment.

This change creates a bit of a noisy diff since JSON5 doesn't require
as many quotes as JSON, so submitting it separately.

[JSON5]: https://json5.org