feat(lint-pr-title): load configuration files from our directory (#546)
* test(lint-pr-title): add integration test for separate config

We have an issue currently where separate configuration files are not
being picked up. This test will reproduce this issue so that we can fix
it in the next commit.

* fix(lint-pr-title): load config file from our directory

`commitlint` configs can have an `extends` key which refers to a module.
There's some runtime support for resolving this module, which requires
it to be discoverable either globally or from the current directory
(e.g. via a `node_modules` contained in it). Here we switch to the
latter, by copying the user-supplied config file to our directory.

This is done by adding a new method of creasting temporary files which
are automatically cleaned up when they go out of scope, using
[`AsyncDisposable`][e-r-m].

[e-r-m]: https://github.com/tc39/proposal-explicit-resource-management