chore(renovate): use upstream `customManagers:githubActionsVersions` instead of our own
I just discovered that Renovate has [a built-in `githubActionsVersions`
manager][manager] which is almost the same as the one we've been
maintaining. Let's switch to that and remove our own. The main
difference is that we supported `_VERSION` and `-version`-style
variables and theirs only supports the former. But it's not hard for us
to switch in the only case we use the matcher.

[manager]: https://docs.renovatebot.com/presets-customManagers/#custommanagersgithubactionsversions