feat(drone-sig): add check for fork
Since this reusable workflow is used by OSS projects, we need to be
aware if we're dealing with a fork or not, especially when it comes to
PRs. If we're on a fork, we don't have access to the secret needed to
run the drone ci config signature validation. In this case, we post a
comment to the PR saying we're skipping the check and someone with
proper repo persmissions needs to check/sign the drone ci config
changes.