chore(lint-pr-title): use text lockfile, specify bun version in separate file
Renovate isn't currently updating the Bun version we're using for this
workflow. That's because we give it in `packageManagers` but it's [not a
supported Corepack package manager][package-managers]. [It does, however,
support updating a `.bun-version` file.][renovate-bun-version] We can
switch to putting our version in that and enable the manager. We've
given `1.1.43` as the version even though `1.1.44` is the newest right
now, so we can see if the update is raised properly.

In `1.1.39` a [new, faster, diffable, text-based lockfile format was
introduced][lockfile-format]. We can switch to that too.

[lockfile-format]: https://bun.sh/blog/bun-lock-text-lockfile#cached-bun-install-gets-30-faster
[package-managers]: https://nodejs.org/api/corepack.html#supported-package-managers
[renovate-bun-version]: https://docs.renovatebot.com/modules/manager/bun-version/