fix(zizmor): fix check for new versions
The variable needs to be named `..._VERSION` for the `# renovate:`
comment to work.

Additionally, a fix is needed in the regex used for such comments. It
currently matches too much in YAML files, so lines _after_ the version
are caught and this breaks the logic. Fortunately, the same fix had been
implemented internally for the same reason, so we can copy it here.