fix: workflow runs from forks (#945)
* ci(reusable-zizmor): fix when run from a fork and there are findings

We can't comment in that situation, so don't try to - it makes the
workflow fail. Findings are present in the "Changed Files" tab and the
CI log, so this isn't so bad.

* ci(test-login-to-gar): don't run for forks

Forks can't access secrets, so this isn't going to work. Skip CI there -
it'll run in the merge queue anyway.

* ci(renovete): fix workflow runs from forks, run in merge queue

This needs access to a secret which isn't available in forks, so
don't run there. Add the `merge_group` trigger, so that we do run in the
merge queue after the PR has been queued by a maintainer.

* ci(test-find-pr-for-commit): don't run for forks

I have been trying to fix this but I haven't been able to get it to work
just yet. Right now I need to move on unfortunately. We can come back to
this later.