feat(reusable-zizmor): store default config file separately (#953)
We've got a TODO item saying that storing the config in the YAML
definition of the reusable workflow itself isn't the best. It's hard to
find, edit, and there's no linting since it's considered a string and
not YAML itself.

It's not super obvious how to do this, though. What we want to do is use
a file which is alongside the workflow. So if we're called like:

```yaml
job:
  zizmor:
    uses: grafana/shared-workflows/.github/workflows/reusable-zizmor.yml@abc123
```

...the correct thing to do is to use the file from `abc123` and not
anything else which might be incompatible or change independently of the
caller updating their pointer to us.

GitHub doesn't give us a checkout of the reusable workflow. So we need
to fetch the file from GitHub ourselves. What we need to do that is the
ref (`abc123`). Unfortunately this isn't easily available. For example
it's not in the `${{ github }}` context, or any of the predefined
environment variables.

But one place it _is_ available is [in the OIDC ID
token][oidc-reusable].

Given that, by fetching the `job_workflow_ref` from the ID token, we can
find out our current reference and retrieve the default config file from
the repository at that ref.

This lets us store it separately, rather than having it inline where
it's much harder to edit.

Here we do that - we fetch the ID token, convert it to a SHA if it's a
named ref like `refs/heads/main`, and then fetch `.github/zizmor.yml`
from this repository at that SHA. The file is cached, so we don't have
re-fetch the file every time.

[oidc-reusable]: https://docs.github.com/en/actions/security-for-github-actions/security-hardening-your-deployments/using-openid-connect-with-reusable-workflows