fix(reusable-zizmor): fix runs from forks (#970)
Forks can't get an ID token. We use this method to look up the version
we're called at: get an OIDC ID token and then use the
`job_workflow_ref` claim. We're currently hard failing when this doesn't
work, which means we throw an error for runs from forks.

It would be better to still succeed, and fall back to using the default
config.