Replace hardcoded /opt directory with RUNNER_TEMP
Non-root users are not allowed to write in the /opt directory.
Instead we leave it to github runners to create stuff in a temp directory
that every user can access.

Note that we are allowing zizmor to ignore a github env error.
This is because the RUNNER_TEMP variable is controlled just by GITHUB_ACTIONS and
is not accessible to anyone else