feat(get-latest-workflow-artifact): consider in-progress runs
In some situations a workflow might get triggered that depends on an
artifact while the producing workflow is still in-progress.

There might even be the case that the artifact hasn't been uploaded yet
and so this scenario also allows for 5 retries taking place.

The action now also returns the status of the used workflow as output.