fix(renovate): fix runs from on-forks
In c0944351c68e107649cb9c9a86204b400b607564 we fixed the Renovate job to
be skipped for PRs from forks. But at the same time we broke non-fork
runs. The `if` condition was looking for `github,evnet.pull_request`,
and of course that's not set for any other event.

Fix it so that we always run for non-PR events. The `on:` conditions
take care of us there.