ci: enhance, fix, run pre-commit (#1033)
I got an unrelated linting error in #1032. This could have been caught
with `pre-commit` - add it here, and fix all the linting errors it shows
right now.

The Go hook doesn't work properly in a monorepo, so it's disabled.

We can run pre-commit in CI. This duplicates many of the checks we were
running already, so the old ones are removed.