fix(renovate): use a dry-run when run from the merge group (#1097)
We've currently re-running Renovate for _any_ run in the merge group,
which is making it push branches whenever anything is queued to be
merged. This is causing way too many runs and unnecessary rebases and CI
runs. We should only run in write mode on a schedule, on an explicit
manual request, or from the main branch if Renovate's config itself
changes