fix: remediate latest zizmor findings, fix supplying zizmor config (#1101)
The current version of Zizmor finds some potential template injection
issues. We can fix these by indirecting via the `env`, or in a couple of
cases by adding ignore comments where we can't really fix the issue.

Something happened to break our config discovery. What we do is download
a default config file from this repo, write it to a temporary file, set
that file's path as `ZIZMOR_CONFIG`, and then pass this as `--config` if
it's set. Possibly as the result of a version bump, Zizmor started
handling `ZIZMOR_CONFIG` differently. An empty string here is treated as
a file to search for, which doesn't work, and so we get errors.

A fix for this last one is to use a different variable name that doesn't
collide with the one Zizmor itself is using. Rename to
`ZIZMOR_CONFIG_PATH` accordingly.