chore(trigger-argo-workflow): use cli/v3 (#1127)
* chore(trigger-argo-workflow): use cli/v3

* Move submit command into proper subcommand

This is necessary because otherwise the extraArgs would interfere with
the rest of the argument parsing. Despite that, this change should not
be breaking since the `submit` command was also necessary in the
previous implementation.

* Add test for root commandline parsing

This should prevent issues with changes down the line with the way that
cli/v* deals with arguments and flags on subcommands.

* Remove not-used check condition