docs: add instructions for using Renovate (#1164)
* docs: add instructions for using Renovate

Our versioning schema doesn't work with Renovate out of the box.
Fortunately it can be configured to handle it fine. Here we add
instructions for users so that they can set their Renovate configuration
up correctly.

* Update config based on the worning preset in grafana-renovate-config

Doesn't link to the actual preset from that repository, since it's not yet public.

* Fix json

* Lint

* prettier --write

---------

Co-authored-by: Dimitris Sotirakis <dimitrios.sotirakis@grafana.com>