feat: support passing 'builder' parameter in push-to-gar-docker (#1568)
* feat: support passing 'builder' parameter in push-to-gar-docker

We are trying to use the [buildkit-cache-dance][1] method to get better
caching of our Go build cache in CI, but that requires having control
over the buildx builder in our pipelines. This commit adds the 'builder'
input to the push-to-gar-docker shared action, and conditionally creates
a new builder if it isn't provided.

[1]: https://github.com/reproducible-containers/buildkit-cache-dance/

* Fix zizmor issues

* Include builder in table of inputs/outputs in README